pyproject.toml
[tool.poetry]
name = "price-formation"
version = "0.1.0"
description = ""
authors = ["Alexander Perov <perov97@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
pillow = "^10.2.0"
coverage = "^7.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
